+++ tests/test_asn1.py
@@ -219,7 +219,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\x30\x08\x02\x01\x01\x04\x03foo'
         assert res_indefinite == b'\x30\x80\x02\x01\x01\x04\x03foo\x00\x00'
 
@@ -246,7 +246,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\x30\x1c\x02\x01\x01\x30\x0a\x04\x03foo\x04\x03bar\x30\x0b\x30\x09\x30\x07\x30\x05\x04\x03boo'
         assert res_indefinite == b'\x30\x80\x02\x01\x01\x30\x80\x04\x03foo\x04\x03bar\x00\x00\x30\x80\x30\x80\x30\x80\x30\x80\x04\x03boo\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00'
 
@@ -258,7 +258,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\x30\x1c\x02\x01\x01\x30\x0a\x04\x03foo\x04\x03bar\x30\x0b\x30\x09\x30\x07\x30\x05\x04\x03boo'
         assert res_indefinite == b'\x30\x80\x02\x01\x01\x30\x80\x04\x03foo\x04\x03bar\x00\x00\x30\x80\x30\x80\x30\x80\x30\x80\x04\x03boo\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00'
 
@@ -273,7 +273,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\x30\x06\x02\x01\x01\x02\x01\x02'
         assert res_indefinite == b'\x30\x80\x02\x01\x01\x02\x01\x02\x00\x00'
 
@@ -288,7 +288,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\x31\x08\x02\x01\x01\x04\x03foo'
         assert res_indefinite == b'\x31\x80\x02\x01\x01\x04\x03foo\x00\x00'
 
@@ -303,7 +303,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\x31\x06\x02\x01\x01\x02\x01\x02'
         assert res_indefinite == b'\x31\x80\x02\x01\x01\x02\x01\x02\x00\x00'
 
@@ -317,7 +317,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\xa1\x03\x02\x01\x01'
         assert res_indefinite == b'\xa1\x80\x02\x01\x01\x00\x00'
 
@@ -331,7 +331,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\x61\x03\x02\x01\x01'
         assert res_indefinite == b'\x61\x80\x02\x01\x01\x00\x00'
 
@@ -345,7 +345,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\xe1\x03\x02\x01\x01'
         assert res_indefinite == b'\xe1\x80\x02\x01\x01\x00\x00'
 
@@ -359,7 +359,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\xff\x83\xff\x7f\x03\x02\x01\x01'
         assert res_indefinite == b'\xff\x83\xff\x7f\x80\x02\x01\x01\x00\x00'
 
@@ -373,7 +373,7 @@ class TestEncoder(object):
         res_default = write_into(asn1.Encoder())
         res_definite = write_into(asn1.Encoder(definite=True))
         res_indefinite = write_into(asn1.Encoder(definite=False))
-        assert res_default == res_indefinite
+        assert res_default == res_definite
         assert res_definite == b'\x30\x08\x02\x01\x01\x04\x03foo'
         assert res_indefinite == b'\x30\x80\x02\x01\x01\x04\x03foo\x00\x00'